+++ packages/core/src/rpc/index.test.ts
@@ -2,7 +2,7 @@ import { setupAnvil, setupCommon } from "@/_test/setup.js";
 import type { Common } from "@/common/common.js";
 import type { Network } from "@/config/networks.js";
 import { beforeEach, expect, test } from "vitest";
-import { createRequestQueue } from "./requestQueue.js";
+import { createRequestQueue } from "./index.js";
 
 beforeEach(setupCommon);
 beforeEach(setupAnvil);
