+++ packages/core/src/sync/transport.ts
@@ -1,6 +1,6 @@
+import type { RequestQueue } from "@/rpc/index.js";
 import type { SyncStore } from "@/sync-store/index.js";
 import { toLowerCase } from "@/utils/lowercase.js";
-import type { RequestQueue } from "@/utils/requestQueue.js";
 import type { Address, Hex, Transport } from "viem";
 import { custom, hexToBigInt, maxUint256 } from "viem";
